some functions

+++ src/mesh/half_edge/indices.rs
@@ -1,32 +1,32 @@
 use std::ops::Deref;
 
-#[derive(Copy, Clone, Debug, Default)]
-pub struct ParentIndex(u32);
+#[derive(Copy, Clone, Debug, Default, PartialEq, PartialOrd)]
+pub struct ParentIndex(usize);
 
 impl Deref for ParentIndex {
-    type Target = u32;
+    type Target = usize;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-#[derive(Copy, Clone, Debug, Default)]
-pub struct HalfEdgeIndex(u32);
+#[derive(Copy, Clone, Debug, Default, PartialEq, PartialOrd)]
+pub struct HalfEdgeIndex(usize);
 
 impl Deref for HalfEdgeIndex {
-    type Target = u32;
+    type Target = usize;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-#[derive(Copy, Clone, Debug, Default)]
-pub struct VertexIndex(u32);
+#[derive(Copy, Clone, Debug, Default, PartialEq, PartialOrd)]
+pub struct VertexIndex(usize);
 
 impl Deref for VertexIndex {
-    type Target = u32;
+    type Target = usize;
 
     fn deref(&self) -> &Self::Target {
         &self.0
